+++ src/app/app-routing.module.ts
@@ -1,8 +1,20 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { AuthGuard } from './core/guards/auth.guard';
 
 
-const routes: Routes = [];
+const routes: Routes = [
+  {
+    path: 'admin',
+    loadChildren: () => import('../app/modules/admin/admin.module').then(m => m.AdminModule),
+    canActivate: [AuthGuard]
+  },
+  {
+    path: '',
+    redirectTo: '/productList',
+    pathMatch: 'prefix'
+  }
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],