+++ src/app/app.component.ts
@@ -1,17 +1,29 @@
-import { AfterViewInit, ElementRef } from '@angular/core';
-import { Component, ViewChild } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
+import { AuthService } from './core/services/auth.service';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.less']
 })
-export class AppComponent implements AfterViewInit{
-  @ViewChild('appTitle') title: ElementRef<HTMLElement>;
+export class AppComponent implements OnInit {
 
-  ngAfterViewInit(){
-    this.title.nativeElement.innerText = 'Angular-Shop';
+  isAdmin: boolean;
 
+  constructor(private authService: AuthService) { 
+    
   }
 
+  ngOnInit(): void {
+    this.isAdmin = this.authService.isAdmin;
+  }
+
+  onActivate($event: any, routerOutlet: RouterOutlet) {
+    console.log('Activated Component', $event, routerOutlet);
+  }
+
+  onDeactivate($event: any, routerOutlet: RouterOutlet) {
+    console.log('Deactivated Component', $event, routerOutlet);
+  }
 }