+++ src/app/models/product.model.ts
@@ -1,8 +1,0 @@
-
-export class ProductModel {
-    id: number;
-    name: string;
-    price: number;
-    rate: number;
-    isAvailable: boolean;
-}