+++ src/app/modules/cart/cart.module.ts
@@ -1,6 +1,14 @@
 import { NgModule } from '@angular/core';
 import { CartListComponent } from './cart-list-component/cart-list-component';
-import { SharedModule } from '../shared/shared.module';
+import { RouterModule, Routes } from '@angular/router';
+import { SharedModule } from 'src/app/core/shared.module';
+
+const routes: Routes = [
+  {
+    path: 'cart',
+    component: CartListComponent
+  }
+];
 
 @NgModule({
   declarations: [
@@ -8,6 +16,7 @@ import { SharedModule } from '../shared/shared.module';
   ],
   imports: [
     SharedModule,
+    RouterModule.forChild(routes)
   ],
   exports: [
     CartListComponent