+++ src/app/modules/cart/services/cart.service.ts
@@ -1,36 +1,43 @@
 import { Injectable } from '@angular/core';
-import { observable, Observable, of } from 'rxjs';
-import { BoughtProductModel } from '../../../models/bought-product.model';
-import { ProductModel } from '../../../models/product.model';
+import { Observable, of } from 'rxjs';
+import { LocalStorageService } from 'src/app/core/services/local-storage.service';
+import { BoughtProductModel } from '../../product/models/bought-product.model';
+import { ProductModel } from '../../product/models/product.model';
+import { ProductService } from '../../product/services/products.service';
 
 @Injectable()
 export class CartService {
+    private key = 'CartService';
     private products: Array<BoughtProductModel> = new Array<BoughtProductModel>();
     private counter = 0;
     private totalAmount = 0;
     private isEmptyList = true;
-    private lastModified = new Date();
+
+    constructor(private localStorageService: LocalStorageService,
+                private productService: ProductService) {
+        this.selectProducts();
+    }
 
     addProduct(product: ProductModel): void {
         if (this.contains(product)) {
-            const boughtProduct = this.getBoughtProductByProduct(product);
+            const boughtProduct = this.getBoughtProductById(product.id);
             boughtProduct.quantity++;
         } else {
             this.products.push(new BoughtProductModel(product, 1));
         }
-        this.updateCartData();
+        this.updateAndSaveCartData();
     }
 
     removeProduct(product: ProductModel): void {
         if (this.contains(product)) {
-            const boughtProduct = this.getBoughtProductByProduct(product);
+            const boughtProduct = this.getBoughtProductById(product.id);
             if (boughtProduct.quantity === 1) {
                 const index = this.products.indexOf(boughtProduct);
                 this.products.splice(index, 1);
             } else {
                 boughtProduct.quantity--;
             }
-            this.updateCartData();
+            this.updateAndSaveCartData();
         }
     }
 
@@ -40,7 +47,7 @@ export class CartService {
         }
 
         if (this.contains(product)) {
-            const boughtProduct = this.getBoughtProductByProduct(product);
+            const boughtProduct = this.getBoughtProductById(product.id);
 
             if (quantity === 0) {
                 const index = this.products.indexOf(boughtProduct);
@@ -50,7 +57,7 @@ export class CartService {
                 boughtProduct.quantity = quantity;
             }
         }
-        this.updateCartData();
+        this.updateAndSaveCartData();
     }
 
     getProducts(): Observable<Array<BoughtProductModel>> {
@@ -59,12 +66,12 @@ export class CartService {
 
     // наверное вспомагательный приватный метод?
     contains(product: ProductModel): boolean {
-        return this.getBoughtProductByProduct(product) != null;
+        return this.getBoughtProductById(product.id) != null;
     }
 
     removeAllProducts(): void {
         this.products = new Array<BoughtProductModel>();
-        this.updateCartData();
+        this.updateAndSaveCartData();
     }
 
     get totalQuantity(): number {
@@ -79,11 +86,7 @@ export class CartService {
         return this.isEmptyList;
     }
 
-    get lastModifiedDate(): Date {
-        return this.lastModified;
-    }
-
-    private updateCartData(): void {
+    private updateAndSaveCartData(): void {
         this.counter = 0;
         this.totalAmount = 0;
         this.products.forEach(element => {
@@ -91,20 +94,31 @@ export class CartService {
             this.totalAmount += element.quantity * element.product.price;
         });
 
-        if (this.counter === 0){
+        if (this.counter === 0) {
             this.isEmptyList = true;
-        }else{
+        } else {
             this.isEmptyList = false;
         }
 
-        this.lastModified = new Date();
+        this.localStorageService.setItem(this.key, this.products);
     }
 
-    private getBoughtProductByProduct(product: ProductModel): BoughtProductModel {
+    private selectProducts(): void {
+        const products = this.localStorageService.getItem(this.key) as Array<BoughtProductModel>;
+        if (products != null) {
+            products.forEach(product => {
+                const boughtProduct = this.productService.getProductById(product.product.id);
+                this.products.push(new BoughtProductModel(boughtProduct, product.quantity));
+            });
+            this.updateAndSaveCartData();
+        }
+    }
+
+    private getBoughtProductById(id: number): BoughtProductModel {
         let boughtProduct = null;
 
         this.products.forEach(x => {
-            if (x.product === product) {
+            if (x.product.id === id) {
                 boughtProduct = x;
             }
         });