+++ src/app/modules/first/first.module.ts
@@ -1,11 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { GeneratorService } from 'src/app/core/services/generator';
 import { FirstComponent } from './first-component/first.component';
-
-function generatorFactory() {
-  return new GeneratorService();
-}
 
 
 @NgModule({