+++ src/app/modules/product/product-list-component/product-list-component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
 import { ProductService } from 'src/app/modules/product/services/products.service';
-import { ProductModel } from 'src/app/models/product.model';
-import { CartService } from 'src/app/modules/cart/services/cart.service';
+import { ProductModel } from 'src/app/modules/product/models/product.model';
+import { environment } from 'src/environments/environment';
+import { AuthService } from 'src/app/core/services/auth.service';
 
 @Component({
   selector: 'app-product-list-component',
@@ -11,22 +12,15 @@ import { CartService } from 'src/app/modules/cart/services/cart.service';
 })
 export class ProductListComponent implements OnInit {
 
+  isAdmin: boolean;
   products$: Observable<Array<ProductModel>>;
 
-  constructor(
-    private productService: ProductService,
-    private cartService: CartService) { }
+  constructor(private productService: ProductService,
+        private authService: AuthService) {
+  }
 
   ngOnInit(): void {
+    this.isAdmin = this.authService.isAdmin;
     this.products$ = this.productService.getProducts();
   }
-
-  onBuy(product: ProductModel) {
-    this.cartService.addProduct(product);
-  }
-
-  onRemove(product: ProductModel) {
-    this.cartService.removeProduct(product);
-  }
-
 }