+++ src/app/modules/product/services/products.service.ts
@@ -1,62 +1,47 @@
 import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
-import { ProductModel } from 'src/app/models/product.model';
-
+import { ProductModel } from 'src/app/modules/product/models/product.model';
 
 @Injectable()
 export class ProductService {
 
+    private description = 'Дисплей Super Retina XDR 6,1 дюйма '
+        + 'Неймовірно міцна передня панель Ceramic Shield'
+        + 'A14 Bionic, найшвидший процесор iPhone'
+        + 'Передова система двох камер 12 Мп: надширококутна і ширококутна; Нічний режим, технологія Deep Fusion, режим Smart HDR 3, знімання HDR-відео 4K у стандарті Dolby Vision'
+        + 'Фронтальна камера TrueDepth 12 Мп: Нічний режим, знімання HDR-відео 4K у стандарті Dolby Vision'
+        + 'Надійний захист від води (IP68)4';
+
+    private products: Array<ProductModel> = [
+        new ProductModel(1, 'Apple iPhone Xs 128GB White', 800, 4.5, 'assets/images/iphoneXs.jpg', this.description),
+        new ProductModel(2, 'Apple iPhone 12 128GB Blue', 1000, 4.2, 'assets/images/iphone12.jpg', this.description),
+        new ProductModel(3, 'Apple iPhone 12 128GB Black', 900, 4.3, 'assets/images/iphoneXr.jpg', this.description),
+        new ProductModel(4, 'Apple iPhone 11 Pro 128GB Gray', 1000, 4.3, 'assets/images/iphone11Pro.jpg', this.description),
+        new ProductModel(5, 'Apple iPhone 8 128GB White', 500, 4.3, 'assets/images/iphone8.jpg', this.description)
+    ];
+
     getProducts(): Observable<Array<ProductModel>> {
-        const product = new ProductModel();
-        product.id = 1234;
-        product.name = 'Car';
-        product.price = 100;
-        product.rate = 3;
-        product.isAvailable = true;
+        return of(this.products);
+    }
 
-        const product1 = new ProductModel();
-        product1.id = 2345;
-        product1.name = 'Bus';
-        product1.price = 200;
-        product1.rate = 2;
-        product1.isAvailable = true;
+    getProductById(id: number): ProductModel {
+        const index = this.products.findIndex(x => x.id === id);
+        return this.products[index];
+    }
 
-        const product2 = new ProductModel();
-        product2.id = 3456;
-        product2.name = 'Computer';
-        product2.price = 300;
-        product2.rate = 5;
-        product2.isAvailable = false;
+    updateProduct(id: number, product: ProductModel): void {
+        const currentProduct = this.getProductById(id);
 
-        const product3 = new ProductModel();
-        product3.id = 4567;
-        product3.name = 'Plane';
-        product3.price = 250;
-        product3.rate = 4;
-        product3.isAvailable = true;
+        currentProduct.name = product.name;
+        currentProduct.price = product.price;
+        currentProduct.rate = product.rate;
+        currentProduct.description = product.description;
+    }
 
-        const product4 = new ProductModel();
-        product4.id = 5678;
-        product4.name = 'Cat';
-        product4.price = 50;
-        product4.rate = 4.2;
-        product4.isAvailable = true;
-
-        const product5 = new ProductModel();
-        product5.id = 6789;
-        product5.name = 'Dog';
-        product5.price = 75;
-        product5.rate = 4.5;
-        product5.isAvailable = true;
-
-
-        const arr = new Array<ProductModel>();
-        arr.push(product);
-        arr.push(product1);
-        arr.push(product2);
-        arr.push(product3);
-        arr.push(product4);
-        arr.push(product5);
-        return of(arr);
+    addProduct(product: ProductModel): number {
+        const id = this.products.length + 1;
+        product.id = id;
+        this.products.push(product);
+        return id;
     }
 }