+++ changelog.md
@@ -13,3 +13,13 @@ Pipes
 1. realized OrderByPipe
 2. used default pieps in CartListComponent such as : async, uppercase, currency
 3. modified cartService to return Observable 
+
+Routing
+1. implemented routing for product list
+2. created page for product description
+3. implemented routing for cart
+4. added components orders
+5. implemented admin panel with CanActivate guard
+6. added resolve guard to product-form component
+7. applied lazy loading to admin module
+8. changed cart service to store cart items in the local storage